Rename profile view to profile settings view
