+++ src/constants/profile.js
@@ -10,7 +10,7 @@ const PROFILE_IMAGE_UPDATE_REQUEST = 'PROFILE_IMAGE_UPDATE_REQUEST';
 const PROFILE_IMAGE_UPDATE_SUCCESS = 'PROFILE_IMAGE_UPDATE_SUCCESS';
 const PROFILE_IMAGE_UPDATE_FAILURE = 'PROFILE_IMAGE_UPDATE_FAILURE';
 
-const PROFILE_OPEN = 'PROFILE_OPEN';
+const PROFILE_SETTINGS_OPEN = 'PROFILE_OPEN';
 
 const PROFILE_STATUS_IDLE = 'STATUS_IDLE';
 const PROFILE_STATUS_BUSY = 'STATUS_BUSY';
@@ -30,7 +30,7 @@ export {
   PROFILE_IMAGE_UPDATE_SUCCESS,
   PROFILE_IMAGE_UPDATE_FAILURE,
 
-  PROFILE_OPEN,
+  PROFILE_SETTINGS_OPEN,
 
   PROFILE_STATUS_IDLE,
   PROFILE_STATUS_BUSY,
