+++ src/constants/routes.js
@@ -36,7 +36,7 @@ const ROUTE_SCHEDULED_CALL_ACTIVITY = 'ROUTE_SCHEDULED_CALL_ACTIVITY';
 
 const ROUTE_CREATE_CALL_NOTES = 'ROUTE_CREATE_CALL_NOTES';
 
-const ROUTE_PROFILE = 'ROUTE_PROFILE';
+const ROUTE_PROFILE_SETTINGS = 'ROUTE_PROFILE_SETTINGS';
 
 const STACK_LANDING = 'STACK_LANDING';
 const STACK_ACTIVITIES = 'STACK_ACTIVITIES';
@@ -93,7 +93,7 @@ export {
 
   ROUTE_CREATE_CALL_NOTES,
 
-  ROUTE_PROFILE,
+  ROUTE_PROFILE_SETTINGS,
 
   STACK_LANDING,
   STACK_ACTIVITIES,
